fix(ci): python-full-coverage — split oversized validation_setup.py into validation_command_probe
The PR's validate-toolchain probe work grew src/awf/runtime/validation_setup.py
to 1687 lines, tripping the first-party maintainability guardrail
(test_first_party_code_files_stay_under_line_limit, MAX_FIRST_PARTY_FILE_LINES
= 1500), which failed the python-coverage-shards (8) job and the
python-full-coverage / ci-required rollup.

Extract the self-contained validate-command shell-probe cluster — plus the
shared shell-tokenization primitives it consumes (_shell_tokens,
_ENV_ASSIGNMENT_RE, _first_non_assignment_token_index) — into a new sibling
module awf.runtime.validation_command_probe. validation_setup re-imports the
façade names (validate_command_probe_targets, _leading_executable(s),
_shell_tokens, _first_non_assignment_token_index) as explicit re-exports so the
awf.runtime.validation barrel and validation_runner keep working unchanged.

Pure decomposition, no behavior change: validation_setup drops to 1021 lines
and the new module is 704 lines, both under the limit. Behavior is covered by
the existing probe/runner tests (2282 passing) which import via the barrel.
